Export CreateProjectListPage and NamespacedPage in @openshift-console/dynamic-plugin-sdk-internal
