+++ frontend/packages/console-dynamic-plugin-sdk/src/extensions/console-types.ts
@@ -57,6 +57,16 @@ export type K8sResourceCommon = {
   metadata?: ObjectMetadata;
 };
 
+// Generic, unknown kind. Avoid when possible since it allows any key in spec
+// or status, weakening type checking.
+export type K8sResourceKind = K8sResourceCommon & {
+  spec?: {
+    [key: string]: any;
+  };
+  status?: { [key: string]: any };
+  data?: { [key: string]: any };
+};
+
 export type K8sVerb =
   | 'create'
   | 'get'
