+++ frontend/packages/console-shared/src/components/projects/NamespacedPage.tsx
@@ -1,23 +1,13 @@
 import * as React from 'react';
 import * as cx from 'classnames';
+import {
+  NamespacedPageProps,
+  NamespacedPageVariants,
+} from '@console/dynamic-plugin-sdk/src/api/internal-types';
 import { NamespaceBar } from '@console/internal/components/namespace';
 import NamespaceBarApplicationSelector from '@console/topology/src/components/dropdowns/NamespaceBarApplicationSelector';
 
 import './NamespacedPage.scss';
-
-export enum NamespacedPageVariants {
-  light = 'light',
-  default = 'default',
-}
-
-export interface NamespacedPageProps {
-  disabled?: boolean;
-  hideProjects?: boolean;
-  hideApplications?: boolean;
-  onNamespaceChange?: (newNamespace: string) => void;
-  variant?: NamespacedPageVariants;
-  toolbar?: React.ReactNode;
-}
 
 const NamespacedPage: React.FC<NamespacedPageProps> = ({
   children,
