+++ frontend/public/module/k8s/types.ts
@@ -5,6 +5,7 @@ import {
   ObjectMetadata,
   K8sResourceCommon,
   K8sVerb,
+  K8sResourceKind,
 } from '@console/dynamic-plugin-sdk/src/extensions/console-types';
 import { EventInvolvedObject } from './event';
 import { Selector, MatchLabels, K8sModel } from '@console/dynamic-plugin-sdk/src/api/common-types';
@@ -55,16 +56,6 @@ export type Toleration = {
   operator: TolerationOperator;
   tolerationSeconds?: number;
   value?: string;
-};
-
-// Generic, unknown kind. Avoid when possible since it allows any key in spec
-// or status, weakening type checking.
-export type K8sResourceKind = K8sResourceCommon & {
-  spec?: {
-    [key: string]: any;
-  };
-  status?: { [key: string]: any };
-  data?: { [key: string]: any };
 };
 
 export type VolumeMount = {
